+++ deltachat/rpcbin.go
@@ -1,0 +1,6 @@
+//go:build !windows
+// +build !windows
+
+package deltachat
+
+const deltachatRpcServerBin = "deltachat-rpc-server"