+++ deltachat/rpcbin_windows.go
@@ -1,0 +1,3 @@
+package deltachat
+
+const deltachatRpcServerBin = "deltachat-rpc-server.exe"